gr/vat: web-service client for Greek VAT checker
Since May 2014, the IT of Greek Ministry of Finance is offering a
web-service for VAT-numbers lookup. It is a SOAP service, intended
for individuals.

The implementation is fairly trivial, but access to the service
requires authentication and signup of the client using the web-page
of "www.gsis.gr".

Code originally drafted by Dimosthenes Fioretos <dfiore@noc.uoa.gr>